requirements.txt
fastapi==0.110.0
uvicorn
python-dotenv
pymongo
pydantic>=2.6
pydantic-settings
email-validator
bcrypt==4.0.1
passlib==1.7.4
python-jose[cryptography]
python-multipart
redis
celery[redis]
